requirements.txt
streamlit
google-generativeai
plotly
pandas
requests